15 feature request skip indexing file if already indexed (#26)
* Basic structure

* Move metadata filtering to FilesReader

This way files that are not needed to be indexed again are ignored at
fist sight and are not temporarly stored in a list. This should be
beneficial for large paths to be indexed, however it also tinkers with
the file reading process that is not really the very same concern.
Seperating both steps (reading and filtering, like already implemented
in the commit before) would simplify the whole chain of steps and thus
making it easier to maintain. Also further changes like forcing a
reindex of all files might be easier with these steps being separated.

* Adding metadata query to embeddingsRepository

* Returning filtering to EmbeddingsWrapper